requirements.txt
aiofiles==23.1.0
aiohttp==3.7.4.post0
altair==4.2.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==3.0.1
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.1a1
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
ci-info==0.3.0
click==8.1.3
cohere==3.2.1
comm==0.1.3
configobj==5.0.8
configparser==6.0.0
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
docx2txt==0.8
entrypoints==0.4
etelemetry==0.3.0
exceptiongroup==1.1.0
executing==1.2.0
extra-streamlit-components==0.1.56
faiss-cpu==1.7.3
Faker==18.13.0
favicon==0.7.0
filelock==3.12.2
flake8==6.0.0
fonttools==4.40.0
frontend==0.0.3
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
htbuilder==0.6.1
httplib2==0.22.0
huggingface-hub==0.16.4
ibm-cloud-sdk-core==3.16.7
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.14.0
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
langchain==0.0.79
looseversion==1.3.0
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==2.1.0
markdownlit==0.0.7
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==0.4.3
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
numpy==1.24.1
openai==0.26.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
pathspec==0.11.0
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
plotly==5.15.0
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==3.20.3
prov==2.0.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
pydeck==0.8.0
pydot==1.4.2
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.7.0
pymdown-extensions==10.0.1
Pympler==1.0.1
PyMuPDF==1.22.5
pyparsing==3.0.9
pypdf==3.3.0
PyPDF2==3.0.1
PyPDF4==1.27.0
pyrsistent==0.19.3
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyxnat==1.6
PyYAML==6.0
pyzmq==25.1.0
rdflib==6.3.2
regex==2023.6.3
requests==2.31.0
rich==13.2.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.46
st-annotated-text==4.0.0
stack-data==0.6.2
starlette==0.30.0
streamlit==1.17.0
streamlit-authenticator==0.2.2
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-elements==0.1.0
streamlit-embedcode==0.1.2
streamlit-extras==0.2.7
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.5
streamlit-keyup==0.2.0
streamlit-nested-layout==0.1.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tenacity==8.2.0
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
traits==6.3.2
transformers==4.30.2
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.23.1
validators==0.20.0
wcwidth==0.2.6
yarl==1.8.2
zipp==3.11.0
